# [Do not merge] Benchmark hot path cumulative vs. delta