@mic-max 2021-11-04
Unrelated to this specific change but I feel like interpreting these benchmark results from the text table isn't very easy. I haven't actually run any benchmarks locally myself yet but am wondering if there's some graph generated that includes the mean, error, and std that we could include an image or svg of in benchmark related PRs. Or perhaps even some GitHub bot that will parse the text tables we include in a PR as a comment or in the description and reply with the graphical image. Or maybe running the benchmarks could be a GitHub action that generates a file URL in its output that if people wanted to check that they could (not sure how consistent the performance of running a GitHub action is though..)

Sorry for the rambling just wanted to share some of my thoughts 🙂

PS: [BenchmarkDotNet RPlotExporter](https://benchmarkdotnet.org/articles/overview.html#exporters) Maybe we could use these 👀