@alanwest 2021-11-08
> Sorry for the rambling just wanted to share some of my thoughts 🙂

😆 

> PS: [BenchmarkDotNet RPlotExporter](https://benchmarkdotnet.org/articles/overview.html#exporters) Maybe we could use these 👀

Cool, yea I'll investigate these visualizations and try them out in the future.

Closing this PR for now.